definitions/raster-calculations/containers/download-dem/requirements.txt
python-decouple
requests